requirements.txt
wget
nemo_toolkit[asr]==2.0.0rc0
nltk
faster-whisper>=1.1.0

git+https://github.com/MahmoudAshraf97/demucs.git
git+https://github.com/oliverguhr/deepmultilingualpunctuation.git
git+https://github.com/MahmoudAshraf97/ctc-forced-aligner.git

huggingface_hub<0.24
numpy<2

openai>=1.0.0
openai-whisper==20240930

python-dotenv>=0.21
pydub>=0.25.1

torch==2.6.0
torchaudio==2.6.0
ffmpeg-python>=0.2.0
